crawler/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.4

# Web Crawling
playwright==1.44.0
beautifulsoup4==4.12.3
lxml==5.2.2

# Knowledge Base / AI
chromadb==0.5.0
sentence-transformers==2.7.0
langchain==0.2.5
langchain-community==0.2.5
langchain-text-splitters==0.2.2
torch==2.3.1
tqdm==4.66.4
transformers==4.40.2

# Testing
pytest==8.2.2
pytest-cov==5.0.0